[Engine3D] Item alignment precision to be float level
Why: Item alignment precision to be float level.
How: Item alignment precision to be float level.
Verify steps: N/A
Possible effects: none

+++ src/min3d/component/ScrollContainer.java
@@ -91,9 +91,9 @@ public class ScrollContainer extends Object3dContainer {
             mAlignment = size;
             if (mInit) {
                 if (mMode == CustomScroller.Mode.X) {
-                    mScroller.setItemSize((int) (mAlignment / mRatio));
+                    mScroller.setItemSize(mAlignment / mRatio);
                 } else {
-                    mScroller.setItemSize((int) (mAlignment / mRatio));
+                    mScroller.setItemSize(mAlignment / mRatio);
                 }
             } else {
                 requestLayout();
